engine: change RenderAPI's RenderGetParm return type to intptr_t to insure compatibility with 64-bit

+++ engine/client/client.h
@@ -964,7 +964,7 @@ void CL_ClearAllRemaps( void );
 // cl_render.c
 //
 qboolean R_InitRenderAPI( void );
-int CL_RenderGetParm( const int parm, const int arg, const qboolean checkRef );
+intptr_t CL_RenderGetParm( const int parm, const int arg, const qboolean checkRef );
 lightstyle_t *CL_GetLightStyle( int number );
 int R_FatPVS( const vec3_t org, float radius, byte *visbuffer, qboolean merge, qboolean fullvis );
 const ref_overview_t *GL_GetOverviewParms( void );
